move max memory
